Removed some unnecessary specificity.
There was no need for specific bit sizes for these integers; the
default int will do just fine.

+++ src/gfx.cpp
@@ -316,7 +316,7 @@ void gfx_drawCircle(int xc, int yc, int R, SDL_Surface *PIX, int col)
 
 void gfx_drawRect(SDL_Surface *dest, int x, int y, int w, int h, Uint8 red, Uint8 green, Uint8 blue)
 {
-	SDL_Rect r = {(int16_t)x, (int16_t)y, (uint16_t)w, (uint16_t)h};
+	SDL_Rect r = {x, y, w, h};
 	SDL_FillRect(dest, &r, SDL_MapRGB(screen->format, red, green, blue));
 
 	gfx_drawLine(dest, x, y, x + w, y, SDL_MapRGB(screen->format, 255, 255, 255));
